Extract review GUI dashboard orchestration into visualization_service.
Move render_dashboard_with_thumbnails and session caching out of
simapp_connector so the connector only manages SimApp client lifecycle.
